--- deploy.js
@@ -1,1 +1,0 @@
-console.log('This is where my deploy script would run if I had one.');
